requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
gunicorn==21.2.0
psycopg==3.1.12
python-dotenv==1.0.0
Flask-WTF==1.1.1
Werkzeug==2.3.7
email-validator==2.1.0
Pillow==10.4.0
pytz==2024.1
resend==2.21.0
cryptography==43.0.0
reportlab==4.1.0
pyotp==2.9.0
qrcode[pil]==7.4.2
stripe==10.15.0
